+++ web/srv/api_handlers_test.go
@@ -7,11 +7,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/ghodss/yaml"
 	"github.com/julienschmidt/httprouter"
-	helpers "github.com/linkerd/linkerd2/cli/cmd"
 	"github.com/linkerd/linkerd2/controller/api/public"
-	"github.com/linkerd/linkerd2/controller/gen/apis/serviceprofile/v1alpha1"
 	pb "github.com/linkerd/linkerd2/controller/gen/public"
 )
 
@@ -54,57 +51,5 @@ func TestHandleApiVersion(t *testing.T) {
 		t.Errorf("incorrect api result")
 		t.Errorf("Got: %+v", jsonResult)
 		t.Errorf("Expected to find: %+v", expectedVersionJson)
-	}
-}
-
-func TestHandleConfigDownload(t *testing.T) {
-	mockApiClient := &public.MockApiClient{
-		VersionInfoToReturn: &pb.VersionInfo{
-			GoVersion:      "the best one",
-			BuildDate:      "never",
-			ReleaseVersion: "0.3.3",
-		},
-	}
-	server := FakeServer()
-
-	handler := &handler{
-		render:    server.RenderTemplate,
-		apiClient: mockApiClient,
-	}
-
-	recorder := httptest.NewRecorder()
-	req := httptest.NewRequest("GET", "/profiles/new?service=authors&namespace=booksns", nil)
-
-	handler.handleProfileDownload(recorder, req, httprouter.Params{})
-
-	if recorder.Code != http.StatusOK {
-		t.Errorf("Incorrect StatusCode: %+v", recorder.Code)
-		t.Errorf("Expected              %+v", http.StatusOK)
-	}
-
-	header := http.Header{
-		"Content-Type": []string{
-			"text/yaml",
-		},
-		"Content-Disposition": []string{
-			"attachment; filename='authors-profile.yml'",
-		},
-	}
-	if !reflect.DeepEqual(recorder.Header(), header) {
-		t.Errorf("Incorrect headers: %+v", recorder.Header())
-		t.Errorf("Expected:          %+v", header)
-	}
-
-	var serviceProfile v1alpha1.ServiceProfile
-	err := yaml.Unmarshal(recorder.Body.Bytes(), &serviceProfile)
-	if err != nil {
-		t.Fatalf("Error parsing service profile: %v", err)
-	}
-
-	expectedServiceProfile := helpers.GenServiceProfile("authors", "booksns")
-
-	err = helpers.ServiceProfileYamlEquals(serviceProfile, expectedServiceProfile)
-	if err != nil {
-		t.Fatalf("ServiceProfiles are not equal: %v", err)
 	}
 }
